Consolidate the three opreators into one master operator
