+++ astro-migration-provider/setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 name = astronomer-migration-provider
-version = 0.1.0
-url = https://github.com/astronomer/starship/astronomer
+version = 0.1.1
+url = https://github.com/astronomer/starship/astro-migration-provider
 description = Migration operators to assist moving variables, connections to Astronomer Cloud
 author = Chris Hronek
 long_description = file: README.md
